1. -Implement assert for IOS system. 2. -Use ptrdiff_t to handle pointer cast operation in 64-bit os. 3. -Implement CFilePathTool::Canonical for IOS system.

+++ Components/Component/ComponentInstanceManager.cpp
@@ -346,7 +346,7 @@ void CComponentInstanceManager::ResolveDependency()
             }
             else
             {
-                *(uint32_t*)pDependencyResolver->pVariableAddress = (uint32_t)pComponentToBeLink;
+                *(ptrdiff_t*)pDependencyResolver->pVariableAddress = (ptrdiff_t)pComponentToBeLink;
             }
         }
     }
